### 2023-11-01 #6005 @danielskatz
@jcorbino - thanks for your submission. I see a couple of issues:

1. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and see how to better indicate multiple references at a given point: e.g., `[@author1:2001; @author2:2001]`

2. the JOSS reference check that is part of starting a new pre-review didn't happen for some reason. I'm going to try to make it happen manually.

Command that may be helpful to you going forward are:
* `@editorialbot check references` to check references
* `@editorialbot generate pdf` to make a new PDF

editorialbot commands need to be the first entry in a new comment.
