### 2023-11-01 #6005 @danielskatz
While you fix the multiple references in the .md file and generate a new PDF, I will start looking for an editor.
