### 2023-11-08 #6005 @danielskatz
👋 @jedbrown - a ping to see if you would be able to edit this, as [previously invited](https://github.com/openjournals/joss-reviews/issues/6005#issuecomment-1789633563)
