### 2023-11-09 #6005 @jedbrown
Hi, @jcorbino! It looks like there have been a lot of improvements since this was [previously reviewed](https://github.com/openjournals/joss-reviews/issues/3380). Before we get reviewers, I notice a few things in my pre-review scan. Addressing these now will give reviewers a much better experience and should lead to a faster review.

1. Asking users to modify absolute paths in makefiles is pretty rough in 2023. There are currently five makefiles that hard-code a path for `ARMA` and they use three different paths on the author's computer. Meanwhile, many users will `apt install libarmadillo-dev` or similar and would prefer to use that. If you don't want a "configure" step, I would leave the variable empty and use `$(if $(ARMA),-L$(ARMA)/lib)` so the library flag is only added when specified by the user. This way `make` should work with no makefile modification when armadillo is installed to a system location and `make ARMA=/path/to/armadillo` can be used otherwise.
2. The paper says the examples include Navier-Stokes, but no such example is in the repository.
3. The paper should discuss related methods such as FEEC that also satisfy a discrete de Rham complex (and may be algebraically equivalent in some cases).
4. The build also depends on OpenBLAS, but this is not documented. If you don't need the specific features of OpenBLAS, I would suggest linking `-llapack -lblas`. Flags for a different implementation can be chosen by a make variable passed on the command line.
5. [Amanzi-ATS](https://amanzi.github.io/about) is an example of a production open source solver based on MFD methods.
6. I can't tell from the documentation whether MOLE supports a consistent discretization of tensor diffusivity on non-K-orthogonal grids. This is one of the unique value propositions of MFD, and the reason it is used in Amanzi-ATS (while most subsurface flow packages use two-point flux methods, despite their inconsistency).
7. I haven't finished the install yet, but I recall last time noticing the disturbing property that `elliptic1D` produced a nonsymmetric indefinite matrix with complex eigenvalues. Is that still the case or has that been resolved?
8. The advection examples seem to use forward Euler and leapfrog on a uniform grid with CFL exactly 1, which is sort of cheating because naive upwinding is exact. Could you use a non-uniform mesh or other slightly more discerning setup?
9. Are there any convergence tests or examples that would demonstrate the advantage of MFD over standard discretizations?
