+++ mole_C++/gradient.cpp
@@ -23,7 +23,7 @@ Gradient::Gradient(u16 k, u32 m, Real dx) : sp_mat(m+1, m+2)
                 at(i, i+1) =  1.0;
             }
             // Weights
-            P << 3.0/8.0 << 9.0/8.0 << 1.0 << 9.0/8.0 << 3.0/8.0;
+            P = {3.0/8.0, 9.0/8.0, 1.0, 9.0/8.0, 3.0/8.0};
             break;
         case 4:
             // A
@@ -56,9 +56,8 @@ Gradient::Gradient(u16 k, u32 m, Real dx) : sp_mat(m+1, m+2)
                 at(i, i+2) = -1.0/24.0;
             }
             // Weights
-            P << 1606.0/4535.0 <<  941.0/766.0  << 1384.0/1541.0 << 1371.0/1346.0
-              <<  701.0/700.0  << 1371.0/1346.0 << 1384.0/1541.0 <<  941.0/766.0
-              << 1606.0/4535.0;
+            P = {1606.0/4535.0, 941.0/766.0, 1384.0/1541.0, 1371.0/1346.0, 701.0/700.0,
+                 1371.0/1346.0, 1384.0/1541.0, 941.0/766.0, 1606.0/4535.0};
             break;
         case 6:
             // A
@@ -115,11 +114,9 @@ Gradient::Gradient(u16 k, u32 m, Real dx) : sp_mat(m+1, m+2)
                 at(i, i+3) =   3.0/640.0;
             }
             // Weights
-            P <<  420249.0/1331069.0 << 2590978.0/1863105.0 <<  882762.0/1402249.0
-              << 1677712.0/1359311.0 <<  239985.0/261097.0  <<  664189.0/657734.0
-              <<  756049.0/754729.0  <<  664189.0/657734.0  <<  239985.0/261097.0
-              << 1677712.0/1359311.0 <<  882762.0/1402249.0 << 2590978.0/1863105.0
-              <<  420249.0/1331069.0;
+            P = {420249.0/1331069.0, 2590978.0/1863105.0, 882762.0/1402249.0, 1677712.0/1359311.0,
+                 239985.0/261097.0, 664189.0/657734.0, 756049.0/754729.0, 664189.0/657734.0, 239985.0/261097.0,
+                 1677712.0/1359311.0, 882762.0/1402249.0, 2590978.0/1863105.0, 420249.0/1331069.0};
             break;
         case 8:
             // A
@@ -208,12 +205,11 @@ Gradient::Gradient(u16 k, u32 m, Real dx) : sp_mat(m+1, m+2)
                 at(i, i+4) =    -5.0/7168.0;
             }
             // Weights
-            P <<  267425.0/904736.0  << 2307435.0/1517812.0 <<  847667.0/3066027.0
-              << 4050911.0/2301238.0 <<  498943.0/1084999.0 <<  211042.0/170117.0
-              << 2065895.0/2191686.0 << 1262499.0/1258052.0 << 1314891.0/1312727.0
-              << 1262499.0/1258052.0 << 2065895.0/2191686.0 <<  211042.0/170117.0
-              <<  498943.0/1084999.0 << 4050911.0/2301238.0 <<  847667.0/3066027.0
-              << 2307435.0/1517812.0 <<  267425.0/904736.0;
+            P = {267425.0/904736.0, 2307435.0/1517812.0, 847667.0/3066027.0, 4050911.0/2301238.0,
+                 498943.0/1084999.0, 211042.0/170117.0, 2065895.0/2191686.0, 1262499.0/1258052.0,
+                 1314891.0/1312727.0, 1262499.0/1258052.0, 2065895.0/2191686.0, 211042.0/170117.0,
[…]
