+++ mole_C++/divergence.cpp
@@ -14,7 +14,7 @@ Divergence::Divergence(u16 k, u32 m, Real dx) : sp_mat(m+2, m+1)
                 at(i, i)   =  1.0;
             }
             // Weights
-            Q << 1.0 << 1.0 << 1.0 << 1.0 << 1.0;
+            Q = {1.0, 1.0, 1.0, 1.0, 1.0};
             break;
         case 4:
             // A
@@ -38,9 +38,8 @@ Divergence::Divergence(u16 k, u32 m, Real dx) : sp_mat(m+2, m+1)
             }
             // Weights
             //
-            Q << 2186.0/1943.0 << 2125.0/2828.0 << 1441.0/1240.0 <<  648.0/673.0
-              <<  349.0/350.0  <<  648.0/673.0  << 1441.0/1240.0 << 2125.0/2828.0
-              << 2186.0/1943.0;
+            Q = {2186.0/1943.0, 2125.0/2828.0, 1441.0/1240.0, 648.0/673.0, 349.0/350.0,
+                 648.0/673.0, 1441.0/1240.0, 2125.0/2828.0, 2186.0/1943.0};
             break;
         case 6:
             // A
@@ -83,10 +82,9 @@ Divergence::Divergence(u16 k, u32 m, Real dx) : sp_mat(m+2, m+1)
                 at(i, i+2) =   3.0/640.0;
             }
             // Weights
-            Q << 2383.0/2005.0 <<  929.0/2002.0 <<  887.0/531.0  << 3124.0/5901.0
-              << 1706.0/1457.0 <<  457.0/467.0  << 1057.0/1061.0 <<  457.0/467.0
-              << 1706.0/1457.0 << 3124.0/5901.0 <<  887.0/531.0  <<  929.0/2002.0
-              << 2383.0/2005.0;
+            Q = {2383.0/2005.0, 929.0/2002.0, 887.0/531.0, 3124.0/5901.0, 1706.0/1457.0,
+                 457.0/467.0, 1057.0/1061.0, 457.0/467.0, 1706.0/1457.0, 3124.0/5901.0,
+                 887.0/531.0, 929.0/2002.0, 2383.0/2005.0};
             break;
     }
 

[…]
