+++ JOSS/paper.bib
@@ -74,7 +74,7 @@
   journal = {Matrix Analysis and Applications},
   month = July,
   title = {{A Matrix Analysis Approach to Higher-Order Approximations for Divergence and Gradients Satisfying a Global Conservation Law}},
-  year = 2006,
+  year = 2003,
   volume = 25
 }
 
@@ -190,4 +190,89 @@
   doi = {10.13140/RG.2.2.28307.86561},
   url = {http://dx.doi.org/10.13140/RG.2.2.28307.86561},
   year = {2023}
+}
+
+@book{Arnold,
+  author = {Arnold, Douglas N.},
+  title = {Finite Element Exterior Calculus},
+  publisher = {Society for Industrial and Applied Mathematics},
+  year = {2018},
+  doi = {10.1137/1.9781611975543},
+  address = {Philadelphia, PA},
+  edition = {},
+  URL = {https://epubs.siam.org/doi/abs/10.1137/1.9781611975543},
+  eprint = {https://epubs.siam.org/doi/pdf/10.1137/1.9781611975543}
+}
+
+@book{da_Veiga,
+  title = {The Mimetic Finite Difference Method for Elliptic Problems},
+  ISBN = {9783319026633},
+  url = {http://dx.doi.org/10.1007/978-3-319-02663-3},
+  DOI = {10.1007/978-3-319-02663-3},
+  publisher = {Springer International Publishing},
+  author = {da Veiga, Lourenço Beirão and Lipnikov, Konstantin and Manzini, Gianmarco},
+  year = {2014}
+}
+
+@techreport{Eldred,
+  title = {Differential geometric approaches to momentum-based formulations for fluids [Slides]},
+  author = {Eldred, Christopher and Stewart, Jim},
+  abstractNote = {This SAND report documents CIS Late Start LDRD Project 22-0311, "Differential geometric approaches to momentum-based formulations for fluids". The project primarily developed geometric mechanics formulations for momentum-based descriptions of nonrelativistic fluids, utilizing a differential geometry/exterior calculus treatment of momentum and a space+time splitting. Specifically, the full suite of geometric mechanics formulations (variational/Lagrangian, Lie-Poisson Hamiltonian and Curl-Form Hamiltonian) were developed in terms of exterior calculus using vector-bundle valued differential forms. This was done for a fairly general version of semi-direct product theory sufficient to cover a wide range of both neutral and charged fluid models, including compressible Euler, magnetohydrodynamics and Euler-Maxwell. As a secondary goal, this project also explored the connection between geometric mechanics formulations and the more traditional Godunov form (a hyperbolic system of conservation laws). Unfortunately, this stage did not produce anything particularly interesting, due to unforeseen technical difficulties. There are two publications related to this work currently in preparation, and this work will be presented at SIAM CSE 23, at which the PI is organizing a mini-symposium on geometric mechanics formulations and structure-preserving discretizations for fluids. The logical next step is to utilize the exterior calculus based understanding of momentum coupled with geometric mechanics formulations to develop (novel) structure-preserving discretizations of momentum. This is the main subject of a successful FY23 CIS LDRD "Structure-preserving discretizations for momentum-based formulations of fluids".},
+  doi = {10.2172/1890065},
+  url = {https://www.osti.gov/biblio/1890065},
+  journal = {Sandia National Laboratories},
+  number = ,
+  volume = ,
+  place = {United States},
+  year = {2022},
+  month = {9}
+}
+
+@phdthesis{Hirani,
+  author = {Hirani,Anil N.},
+  year = {2003},
+  title = {Discrete exterior calculus},
+  journal = {ProQuest Dissertations and Theses},
+  pages = {103},
+  note = {Copyright - Database copyright ProQuest LLC; ProQuest does not claim copyright in the individual underlying works; Last updated - 2023-03-03},
[…]
