### 2023-12-14 #6005 @jcorbino
Thank you @jedbrown for the quick feedback! Here are our answers to your questions:

1. Suggestion was incorporated.

2. As you pointed out, there’s no example with a filename “Navier-Stokes.m” or similar, yet Navier-Stokes equations are solved in lock_exchange.m, which solves a Lock-exchange problem for a Boussinesq fluid. In [this article](https://doi.org/10.1016/j.compfluid.2023.105817) (referenced in the draft), MOLE is used to solve NS in various scenarios.

3. **This article focuses on the open-source library that implements mimetic operators as defined in the work of Corbino and Castillo (2020).** The scope of this article is limited to the discussion of the mentioned library, and it does not delve into alternative methods or their efficacy.

Our paper deals with the implementation of a computational library for MD. **From the current draft:**

_“The mathematics are based on the work of (Corbino & Castillo, 2020).”_

_“A previous library (Sanchez et al., 2014) was developed to implement the mimetic operators
presented in (J. E. Castillo & Grone, 2006). This library was only capable of handling dense matrices so it was limited to solving small problems hence its development was stopped. MOLE implements the operators presented in the Corbino and Castillo paper (Corbino & Castillo, 2020).”_

_“MOLE was designed to be an intuitive software package to construct mimetic operators based on (Corbino & Castillo, 2020) method.”_

_“For an overview of mimetic methods of different types, see the book by Castillo and Miranda and the references therein (José E. Castillo & Miranda, 2013).”_

Furthermore, here is a link to a [technical report](https://www.csrc.sdsu.edu/research_reports/CSRCR2023-10.pdf) where different mimetic methods are discussed.

4. Suggestion was incorporated. **We have updated our installation documentation to clarify that OpenBLAS is required**.

5. Indeed, the Amanzi-ATS library is based on MFD. However, MOLE is developed for MD. MOLE is a tool that accelerates the implementation of numerical solutions of PDEs with MD discrete analogs. **This paper is not about comparing the different mimetic methods.** It is about giving access to researchers to a library that implements MD operators.

6. Currently, MOLE provides the user with mimetic operators designed for PDEs expressed through gradient, curl, divergence, and Laplacian operations. However, in an upcoming release, we will introduce a user-friendly mechanism for supplying a tensor, incorporating information such as anisotropy.

7. From finite difference discretization of the Laplacian, one expects negative eigenvalues. For MD, the 1D Laplacian is computed at the cell centers and not at the nodes, like MFD. The MD 1D Laplacian (L), for m cells, is a product of the divergence (D) and the gradient (G). If one computes the eigenvalues of L by `eig(full(lap(k,m >= 2k+1,1))), k = 2,4,6,8,` one gets m negative eigenvalues together with two zero eigenvalues. The two zero eigenvalues come from the first and last zero rows of the divergence operator (since the divergence has a coordinate-free meaning that requires a neighborhood around a point, there is no sense in computing divergence at the boundaries), and since boundary points are not centers they should be discarded.

**On the other hand, Laplace discretizations do not need to be symmetric**. For example, consider the case of a Laplacian on a hexagonal mesh computed by Hirani's discrete exterior calculus. In that case the Laplacian of 0-form f at a vertex `v0` is given by `(1/h^2) (4/3 f(v1) + 4/3 f(v2) + 4/3 f(v3) - 4 f(v0))`, where `v1, v2, v3` are the simplicial complex formed from the "stencil" consisting of the Voronoi dual mesh associated to `v0`.

8. The main point of the examples (particularly the 1D hyperbolic) is to show the user how to ask the library for the mimetic operators (as defined in Corbino & Castillo, 2020) and how to apply them to a field (scalar or vectorial).

9. As mentioned in the introductory paragraphs, **this work is not about MFD**. One can find an example with theoretical and numerical convergence results in the case of MD in [this report](https://www.csrc.sdsu.edu/research_reports/CSRCR2023-05.pdf).

@mdumett here's where the reviewing process for JOSS takes place :+1:
