### 2024-01-26 #6005 @jedbrown
:wave: @victorapm @josemanuel22 @jakelangham :wave:  Hi, would you be available to review this submission for JOSS?
