### 2024-02-17 jcorbino/mole#11 @jakelangham
[joss review] 1D Poisson example
I am looking over your JOSS paper preprint as part of https://github.com/openjournals/joss-reviews/issues/6288. Here are a couple of small things I noticed regarding the MATLAB Poisson solver example given.

- If the example is to be included, it should be possible to read from the text (not the code) exactly what the code snippet will do. This would look something like: defining the problem exactly, giving the right-hand side and boundary conditions and stating that the code approximates the solution to the problem by constructing the corresponding 4th order mimetic operator and directly inverting it.
- Likewise, the figure caption ought to be self-contained, or at least reference an equation where the problem is defined. Note that BVP should be defined since it won't count as a standard abbreviation for this journal.
- The example code doesn't actually reproduce the figure exactly because the analytical solution isn't plotted.
- While a 2D Poisson solver is given in examples_MATLAB/, I don't see this 1D code. I think it should be added there since this is the first example that most new users will likely want to look at.
