### 2024-02-17 jcorbino/mole#12 @jakelangham
[joss review] MATLAB tests
As part of the ongoing reviews at https://github.com/openjournals/joss-reviews/issues/6288. It is a requirement for acceptance into JOSS for your repository to contain testing 'to objectively check the expected functionality of the software'.

I can see a folder marked tests_C++/ which I will look over in due course. Are there also tests available for the MATLAB part of the library?
