### 2024-02-17 jcorbino/mole#13 @jakelangham
[joss review] Algebra between operators defined on different grids
While playing around with the MATLAB component of MOLE (as part of https://github.com/openjournals/joss-reviews/issues/6288), I noticed that it is possible to add or multiply operators that are constructed with respect to different grids.

e.g. grad(2, 10, 1) + grad(2, 10, 2)

It also seems to be possible to add together operators on uniform and non-uniform grids.

I'm not sure what this would mean exactly. Is ever desirable to be able to do this? If not then perhaps the code should throw an error in these cases.
