### 2024-02-19 jcorbino/mole#13 @jcorbino
You're correct @jakelangham . Non-unary operations involving operands constructed over different grids can be performed, and this is certainly not a justified scenario (not that I can think of). I don't think one can handle exceptions of that kind without defining the operators (currently simply sparse matrices) as objects of a class. This will be a fundamental change in the code, **which I'm not opposed to**, but it will certainly require a significant amount of time.
