### 2024-02-19 jcorbino/mole#12 @jcorbino
@jakelangham The tests in `tests_C++` are built and executed at the end of the installation process for the C++ flavor of MOLE. The MATLAB flavor of our library does not count with an equivalent folder. We can create a set of tests (similar to those for the C++ case) that the MATLAB user would be encouraged to execute before using the library. Will that fulfill the requirement?
