### 2024-02-21 jcorbino/mole#13 @jakelangham
Yep, I suspected as much. Perhaps it might be beneficial to the library in the future, especially if there are other reasons why supplying additional class data with the operators turns out to be useful.

In the short term, I would recommend that the documentation is supplied with a warning about this. Of course it's not likely to cause users problems in simple scenarios, but in e.g. a more complex code testing various grid refinements, one could foresee it leading to bugs down the line.
