### 2024-02-21 jcorbino/mole#12 @jakelangham
Hi @jcorbino. Yes, my view would be that there should be MATLAB scripts that are equivalent to the C++ tests (to the extent that is possible) and this is necessary to satisfy the requirement for testing. However, on looking over the C++ tests, I do not believe these are currently sufficient for verifying the correct functionality of the library.

From what I can see, the tests verify some but not all of the claimed properties of MOLE operators as stated in https://github.com/jcorbino/mole/blob/master/CSRC%20Report%20on%20MOLE.pdf and only do so for the simplest constructions (e.g. the first three are 2nd order operators on regular minimal size grid with dx = 1).

A more comprehensive set of tests would look like a set of scripts that iterated over each approximation order and the various different types of supported grids, checking each of the various calculus identities _and_ verifying the convergence of the approximations made (e.g. by applying them to a known function and showing that the computation converges to the corresponding analytical result as dx -> 0).
