### 2024-04-20 #6288 @jcorbino
Hello @jedbrown - Apologies for the delayed response. I'm currently addressing the issues raised by Jake Langham. I think I'll have those ready by the end of this month.

At the same time, I see that a good portion of his checklist has non-checked items (unrelated to the issues he mentioned in the repo), such as the ones in the "**Software paper**" section. I don't know the status of that. I assume he's been busy, and he'll review those once I integrate the suggested modifications.

Thank you for your patience.
