### 2024-04-27 jcorbino/mole#15 @jakelangham
[joss review] Paper comments
Hi,

Here are a few comments on your JOSS paper, as part of https://github.com/openjournals/joss-reviews/issues/6288. With apologies in advance - many are quite pedantic / editorial in nature. A select few are more important but hopefully ought to be easy enough to address.

Summary:

- Phrases like 'high-quality' should be avoided in a scientific paper.
- I believe operator names only warrant capitalisation if they refer to proper nouns - i.e. 'Laplacian' is capitalised, but 'gradient' is not.
- 'These operators (matrices) act on staggered grids (uniform, nonuniform, and curvilinear'. Would it be more accurate to say something to the effect of 'These operators act on functions discretised over staggered grids'?

Mimetic operators:

- 'making them more faithful to the physics'. I can see that in some situations this is true - e.g. for conservation laws. But it's less obvious to me _a priori_ that a mimetic operator would always be more physically faithful than an alternative at the same order of discretisation. Could you either clarify or else qualify (i.e. 'more faithful in certain contexts') this sentence?
- line 21 typo 'as well of more' -> 'as well as more'
- line 23/34 'also contained' -> 'which are also contained'
- line 24. I'm not sure what 'the model' refers to.

Statement of need:

- As suggested by the JOSS review criteria, there ought to be some mention of who the audience for the library is. Is it applied scientists - physicists, engeineers etc, or mainly numerical analysts interested in developing new methods, for example?
- You might consider expanding on this section a little. Some of the higher-order operators you provide would be particularly complicated for a non-specialist to implement, so it is better to have a well-tested library that is free to use. Another advantage is that it is easy for the user to swap between different grids, resolutions and discretisation orders etc.

The library:

- line 53: 'just need' -> 'they just need'.
- Though it's probably obvious to most readers, it could be better to clarify that 'cell centers' refers to the corresponding spatial numerical grid over which the operator is defined.
- 'C++ version of the library only depends on Armadillo'. I have not gotten round to looking properly at the C++ library yet, but doesn't it also require SuperLU & OpenBLAS (according to the README)?
- I think it's nice to give the code example. While the code itself is pretty readable, it would nevertheless be clearer to define the 1D boundary value problem explicitly before giving the code and noting either up front or in the figure caption that the analytic solution is compared against the discrete result.

Concluding remarks:

- line 69: Consider starting a new sentence at 'however ...'
