+++ mole_C++/Makefile
@@ -7,7 +7,13 @@ CXXFLAGS = -O3
 endif
 CXXFLAGS += -fopenmp
 
-INCPATH = -I. $(if $(ARMA), -I$(ARMA)/include)
+ifdef ARMA
+INCPATH = -I. -I$(ARMA)/include
+else
+ifeq (,$(filter clean,$(MAKECMDGOALS)))
+$(error ARMA is not defined. Please define ARMA with the path to the Armadillo library.)
+endif
+endif
 
 ifdef EIGEN
 CXXFLAGS += -DEIGEN
