+++ mole_C++/Makefile
@@ -11,7 +11,7 @@ ifdef ARMA
 INCPATH = -I. -I$(ARMA)/include
 else
 ifeq (,$(filter clean,$(MAKECMDGOALS)))
-$(error ARMA is not defined. Please define ARMA with the path to the Armadillo library.)
+$(error ARMA is not defined. Please define ARMA with the path to the Armadillo library)
 endif
 endif
 
