### 2024-04-27 jcorbino/mole#12 @victorapm
@jcorbino along these lines, I would like to suggest that you add a small quick-start section to the top-level `README.md` file about running examples and tests with your library (MATLAB and C++). The idea is to make it a bit easier for users to start get started :)
