+++ JOSS/paper.bib
@@ -1,12 +1,43 @@
 @article{Corbino,
-  	url = {https://doi.org/10.1016/j.cam.2019.06.042},
-  	author = {{Corbino}, J. and {Castillo}, J. E.},
-  	doi = {10.1016/j.cam.2019.06.042},
-  	journal = {Computational and Applied Mathematics},
-  	month = January,
-  	title = {{High-order mimetic finite-difference operators satisfying the extended Gauss divergence theorem}},
-  	year = 2020,
-    volume = 364
+  url = {https://doi.org/10.1016/j.cam.2019.06.042},
+  author = {{Corbino}, J. and {Castillo}, J. E.},
+  doi = {10.1016/j.cam.2019.06.042},
+  journal = {Computational and Applied Mathematics},
+  month = January,
+  title = {{High-order mimetic finite-difference operators satisfying the extended Gauss divergence theorem}},
+  year = 2020,
+  volume = 364
+}
+
+@article{SLU,
+  author = {Li, Xiaoye S.},
+  title = {An overview of SuperLU: Algorithms, implementation, and user interface},
+  year = {2005},
+  issue_date = {September 2005},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  volume = {31},
+  number = {3},
+  issn = {0098-3500},
+  url = {https://doi.org/10.1145/1089014.1089017},
+  doi = {10.1145/1089014.1089017},
+  abstract = {We give an overview of the algorithms, design philosophy, and implementation techniques in the software SuperLU, for solving sparse unsymmetric linear systems. In particular, we highlight the differences between the sequential SuperLU (including its multithreaded extension) and parallel SuperLU_DIST. These include the numerical pivoting strategy, the ordering strategy for preserving sparsity, the ordering in which the updating tasks are performed, the numerical kernel, and the parallelization strategy. Because of the scalability concern, the parallel code is drastically different from the sequential one. We describe the user interfaces of the libraries, and illustrate how to use the libraries most efficiently depending on some matrix characteristics. Finally, we give some examples of how the solver has been used in large-scale scientific applications, and the performance.},
+  journal = {ACM Trans. Math. Softw.},
+  month = {sep},
+  pages = {302–325},
+  numpages = {24},
+  keywords = {supernodal factorization, scalability, parallelism, distributed-memory computers, Sparse direct solver}
+}
+
+@article{OBLAS,
+  title={OpenBLAS: An Optimized BLAS Library},
+  author={Zhang, Xianyi and Zhao, Qian and Wang, Hang and Zhu, Chao and Liu, Jiangning},
+  journal={ACM Transactions on Mathematical Software (TOMS)},
+  volume={46},
+  number={3},
+  pages={34:1--34:19},
+  year={2020},
+  publisher={ACM}
 }
 
 @article{Jorge,
