### 2024-04-27 jcorbino/mole#15 @jcorbino
Hi @jakelangham ,

Thank you for your valuable feedback. I have carefully integrated all of your suggestions into the latest draft version.

At this point, I'll be shifting gears to address the remaining (software) pending issues.
