### 2024-04-28 jcorbino/mole#15 @jakelangham
@jcorbino Looks great. I noticed a couple of extra minor editing things while looking through the new draft but otherwise happy for this issue to be closed.

- line 20: bring the reference at the end within the sentence
- line 22: decapitalise Biharmonic
- line 83: capitalise 'castillo-grone'
