+++ JOSS/paper.bib
@@ -52,7 +52,7 @@
 }
 
 @article{Abouali,
-  title = {Stability and Performance Analysis of the Castillo-Grone Mimetic Operators in Conjunction with RK3 Time Discretization in Solving Advective Equations},
+  title = {Stability and performance analysis of the {Castillo-Grone} mimetic operators in conjunction with RK3 time discretization in solving advective equations},
   author = {Abouali, Mohammad and Castillo, Jose E},
   journal = {Procedia Computer Science},
   volume = {18},
