### 2024-04-29 jcorbino/mole#11 @jcorbino
@jakelangham I made the following minor changes:

#1 Removed BVP and referred to the problem as 1D Poisson eqn with Robin BCs.
#2 Added comment line to code snippet in the draft with the file's name (elliptic1D.m).
#3 The order of accuracy in the example (.m) file was changed to 4 (it was 6) to align with the paper's consistency and to avoid potential confusion.
#4 I intentionally omitted the analytical solution's plotting since it wasn't interesting for the snippet. If this can be confusing, I'll gladly add the line that plots the analytical or remove the analytical from the picture.

Please let me know if these changes are enough to fulfill the requirements in this issue so I can proceed to close it.

Thank you,
