### 2024-04-29 jcorbino/mole#11 @jakelangham
@jcorbino Looks good to me. However, I don't see the filename elliptic1D.m referenced in the new draft. Perhaps it was left out unintentionally (unless I've just missed it)?
