### 2024-04-29 jcorbino/mole#11 @jcorbino
@jakelangham My mistake. I forgot to regenerate the PDF using the editorialbot generate pdf command. I observed the changes in the artifact produced by the workflow. I simply added a comment line to the snippet: % File: elliptic1D.m

![capture](https://github.com/jcorbino/mole/assets/24197859/bf56f37f-bb33-49c0-ab6c-b94fc51ccdb9)
