### 2024-04-29 jcorbino/mole#13 @jcorbino
@jakelangham - I added this to the README file:

**NOTE:** Performing non-unary operations involving operands constructed over different grids may lead to unexpected results. While MOLE currently allows such operations without throwing errors, users must exercise caution when manipulating operators across different grids.

And I also added that note to the manual.

I think the only pending issue now is this one: [https://github.com/jcorbino/mole/issues/12](https://github.com/jcorbino/mole/issues/12)
