### 2024-05-05 jcorbino/mole#16 @jakelangham
[JOSS review] C++ Documentation
As part of the ongoing reviews at https://github.com/openjournals/joss-reviews/issues/6288, I have a query about documentation for MOLE. I can see that there is some documentation for the MATLAB part of the library (in the form of the quick start guide), which I was able to follow (though it does not cover everything, as far as I can tell).

Is there equivalent documentation available for the C++ library? The auto-generated doxygen html files cannot really be counted as documentation because they do not specify what any of the functions do. (Compare for example, to the API docs of Armadillo https://arma.sourceforge.net/docs.html where every function, operator and so on are fully specified, often with usage examples.)

I found that I could only get to grips with what the given examples were doing by reading the code for the library directly. An ordinary user that just wants to set up and solve a problem using MOLE ought not to have to do this.
