### 2024-05-05 jcorbino/mole#17 @jakelangham
[JOSS review] Community guidelines
This issue is part of the ongoing reviews at https://github.com/openjournals/joss-reviews/issues/6288,

I am trying to check if there are '... clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support'.

I can see that there is a CONTRIBUTING.md file which could count for point 1). However, I would be more comfortable ticking off this bullet point if there were a section on the main page (README.md) that briefly pointed any users in the right direction for (1-3).
