### 2024-05-05 #6288 @jakelangham
@jedbrown @jcorbino I have finished working my way through the review bullet points. From what I can see, MOLE is a nice piece of software that should be useful to numerical analysts and any scientists working in applications that would particularly benefit from solving PDEs via mimetic operators.

I have left some points unchecked for now, which relate to issues raised in the repository's tracker. Some of these are  substantial. Nevertheless, I would recommend acceptance on their resolution.

Finally, I note that the C++ integration tests are failing at the moment, according to github. Looks like an easy issue, but suppose this should be fixed before any decision on publication!
