+++ mole_C++/Makefile
@@ -11,7 +11,7 @@ ifdef ARMA
 INCPATH = -I. -I$(ARMA)/include
 else
 ifeq (,$(filter clean,$(MAKECMDGOALS)))
-$(error ARMA is not defined. Please define ARMA with the path to the Armadillo library)
+$(warning The path to Armadillo's library was not provided. Will look for it at the standard system directories.)
 endif
 endif
 
