+++ CONTRIBUTING.md
@@ -1,2 +1,22 @@
-First off, thank you for considering contributing to MOLE. It’s people like you that make MOLE such a great library. 
-If you want to contribute to this project, please shoot an email to johnnycorbino@gmail.com. Thanks! =)
+# Contributing to MOLE
+
+Thank you for considering contributing to MOLE! We appreciate your interest and support. Here are some guidelines to help you get started:
+
+## 1. Contributing to MOLE
+
+- **Fork the Repository**: Start by forking the MOLE repository to your GitHub account.
+- **Create a Branch**: Create a new branch for your feature or bug fix.
+- **Make Your Changes**: Implement your changes and open a pull request with a clear description.
+
+## 2. Reporting Issues or Problems
+
+If you encounter any issues or bugs, please report them by following these steps:
+
+- **Check for Existing Issues**: Before creating a new issue, please check the [issue tracker](https://github.com/jcorbino/mole/issues) to see if the problem has already been reported.
+- **Create a New Issue**: If the issue has not been reported, create a new issue with a clear and descriptive title.
+
+## 3. Seeking Support
+
+For support requests or general questions, please feel free to reach out to us via email at [johnnycorbino@gmail.com](mailto:johnnycorbino@gmail.com).
+
+Thank you for your contributions and support!
