### 2024-05-18 2654ee0
Added reference to Community Guidelines on README
