+++ mole_C++/gradient.cpp
@@ -1,286 +1,285 @@
 #include "gradient.h"
 
 // 1-D Constructor
-Gradient::Gradient(u16 k, u32 m, Real dx) : sp_mat(m+1, m+2)
-{
-    assert(!(k%2));
-    assert(k > 1 && k < 9);
-    assert(m >= 2*k);
+Gradient::Gradient(u16 k, u32 m, Real dx) : sp_mat(m + 1, m + 2) {
+  assert(!(k % 2));
+  assert(k > 1 && k < 9);
+  assert(m >= 2 * k);
 
-    switch (k) {
-        case 2:
-            // A
-            at(0, 0) = -8.0/3.0;
-            at(0, 1) =  3.0;
-            at(0, 2) = -1.0/3.0;
-            // A'
-            at(m, m+1) =  8.0/3.0;
-            at(m, m)   = -3.0;
-            at(m, m-1) =  1.0/3.0;
-            // Middle
-            for (u32 i = 1; i < m; i++) {
-                at(i, i)   = -1.0;
-                at(i, i+1) =  1.0;
-            }
-            // Weights
-            P = {3.0/8.0, 9.0/8.0, 1.0, 9.0/8.0, 3.0/8.0};
-            break;
-        case 4:
-            // A
-            at(0, 0) = -352.0/105.0;
-            at(0, 1) =   35.0/8.0;
-            at(0, 2) =  -35.0/24.0;
-            at(0, 3) =   21.0/40.0;
-            at(0, 4) =   -5.0/56.0;
-            at(1, 0) =   16.0/105.0;
-            at(1, 1) =  -31.0/24.0;
-            at(1, 2) =   29.0/24.0;
-            at(1, 3) =   -3.0/40.0;
-            at(1, 4) =    1.0/168.0;
-            // A'
-            at(m, m+1)   = 352.0/105.0;
-            at(m, m)     = -35.0/8.0;
-            at(m, m-1)   =  35.0/24.0;
-            at(m, m-2)   = -21.0/40.0;
-            at(m, m-3)   =   5.0/56.0;
-            at(m-1, m+1) = -16.0/105.0;
-            at(m-1, m)   =  31.0/24.0;
-            at(m-1, m-1) = -29.0/24.0;
-            at(m-1, m-2) =   3.0/40.0;
-            at(m-1, m-3) =  -1.0/168.0;
-            // Middle
-            for (u32 i = 2; i < m-1; i++) {
-                at(i, i-1) =  1.0/24.0;
-                at(i, i)   = -9.0/8.0;
-                at(i, i+1) =  9.0/8.0;
-                at(i, i+2) = -1.0/24.0;
-            }
-            // Weights
-            P = {1606.0/4535.0, 941.0/766.0, 1384.0/1541.0, 1371.0/1346.0, 701.0/700.0,
-                 1371.0/1346.0, 1384.0/1541.0, 941.0/766.0, 1606.0/4535.0};
-            break;
-        case 6:
-            // A
-            at(0, 0) = -13016.0/3465.0;
-            at(0, 1) =    693.0/128.0;
-            at(0, 2) =   -385.0/128.0;
-            at(0, 3) =    693.0/320.0;
-            at(0, 4) =   -495.0/448.0;
-            at(0, 5) =    385.0/1152.0;
-            at(0, 6) =    -63.0/1408.0;
[…]
