+++ tests_C++/test2.cpp
@@ -2,21 +2,22 @@
  * Nullity test of Gradient operator
  */
 
-#include <iostream>
 #include "mole.h"
+#include <iostream>
 
 int main() {
-    int k = 2;
-    int m = 2*k;
-    Real dx = 1;
-    Real tol = 1e-12;
-    
-    Gradient G(k, m, dx);
-    vec field(m+2, fill::ones);
-    
-    vec sol = G*field;
-    
-    norm(sol) < tol ? cout << "\033[1;32mTest PASSED!\033[0m\n" : cout << "\033[1;31mTest FAILED!\033[0m\n";
-    
-    return 0;
+  int k = 2;
+  int m = 2 * k;
+  Real dx = 1;
+  Real tol = 1e-12;
+
+  Gradient G(k, m, dx);
+  vec field(m + 2, fill::ones);
+
+  vec sol = G * field;
+
+  norm(sol) < tol ? cout << "\033[1;32mTest PASSED!\033[0m\n"
+                  : cout << "\033[1;31mTest FAILED!\033[0m\n";
+
+  return 0;
 }
