+++ mole_C++/divergence.cpp
@@ -1,161 +1,158 @@
 #include "divergence.h"
 
 // 1-D Constructor
-Divergence::Divergence(u16 k, u32 m, Real dx) : sp_mat(m+2, m+1)
-{
-    assert(!(k%2));
-    assert(k > 1 && k < 7);
-    assert(m > 2*k);
+Divergence::Divergence(u16 k, u32 m, Real dx) : sp_mat(m + 2, m + 1) {
+  assert(!(k % 2));
+  assert(k > 1 && k < 7);
+  assert(m > 2 * k);
 
-    switch (k) {
-        case 2:
-            for (u32 i = 1; i < m+1; i++) {
-                at(i, i-1) = -1.0;
-                at(i, i)   =  1.0;
-            }
-            // Weights
-            Q = {1.0, 1.0, 1.0, 1.0, 1.0};
-            break;
-        case 4:
-            // A
-            at(1, 0) = -11.0/12.0;
-            at(1, 1) =  17.0/24.0;
-            at(1, 2) =   3.0/8.0;
-            at(1, 3) =  -5.0/24.0;
-            at(1, 4) =   1.0/24.0;
-            // A'
-            at(m, m)   =  11.0/12.0;
-            at(m, m-1) = -17.0/24.0;
-            at(m, m-2) =  -3.0/8.0;
-            at(m, m-3) =   5.0/24.0;
-            at(m, m-4) =  -1.0/24.0;
-            // Middle
-            for (u32 i = 2; i < m; i++) {
-                at(i, i-2) =  1.0/24.0;
-                at(i, i-1) = -9.0/8.0;
-                at(i, i)   =  9.0/8.0;
-                at(i, i+1) = -1.0/24.0;
-            }
-            // Weights
-            //
-            Q = {2186.0/1943.0, 2125.0/2828.0, 1441.0/1240.0, 648.0/673.0, 349.0/350.0,
-                 648.0/673.0, 1441.0/1240.0, 2125.0/2828.0, 2186.0/1943.0};
-            break;
-        case 6:
-            // A
-            at(1, 0) = -1627.0/1920.0;
-            at(1, 1) =   211.0/640.0;
-            at(1, 2) =    59.0/48.0;
-            at(1, 3) =  -235.0/192.0;
-            at(1, 4) =    91.0/128.0;
-            at(1, 5) =  -443.0/1920.0;
-            at(1, 6) =    31.0/960.0;
-            at(2, 0) =    31.0/960.0;
-            at(2, 1) =  -687.0/640.0;
-            at(2, 2) =   129.0/128.0;
-            at(2, 3) =    19.0/192.0;
-            at(2, 4) =    -3.0/32.0;
-            at(2, 5) =    21.0/640.0;
-            at(2, 6) =    -3.0/640.0;
-            // A'
-            at(m, m)     = 1627.0/1920.0;
-            at(m, m-1)   = -211.0/640.0;
-            at(m, m-2)   =  -59.0/48.0;
-            at(m, m-3)   =  235.0/192.0;
-            at(m, m-4)   =  -91.0/128.0;
-            at(m, m-5)   =  443.0/1920.0;
-            at(m, m-6)   =  -31.0/960.0;
[…]

[…]
