### 2024-05-26 0916872
Added minimal tests for MATLAB version
