+++ tests_C++/Makefile
@@ -9,7 +9,7 @@ ifdef ARMA
 INCPATH += -I. -I$(ARMA)/include
 else
 ifeq (,$(filter clean,$(MAKECMDGOALS)))
-$(warning The path to Armadillo's library was not provided. Will look for it at the standard system directories.)
+$(warning The path to Armadillo's library was not provided. I will look for it in the standard system directories.)
 endif
 endif
 
