### 2024-05-27 03777af
Addressing Victor's suggestion on documentation of tests and examples in README file
