### 2024-05-31 287e588
Expanded tests and added quick guide for C++
