+++ tests_C++/Makefile
@@ -4,6 +4,7 @@ CXX ?= g++
 
 INCPATH = -I$(MOLE)
 CXXFLAGS = -O3 -fopenmp -DARMA_DONT_USE_WRAPPER
+LIBS = -L$(MOLE) -lmole -lopenblas
 
 ifdef ARMA
 INCPATH += -I. -I$(ARMA)/include
@@ -13,9 +14,18 @@ $(warning The path to Armadillo's library was not provided. I will look for it i
 endif
 endif
 
-LIBS = -L$(MOLE) -lmole -lopenblas
+ifdef EIGEN
+CXXFLAGS += -DEIGEN
+INCPATH += -I$(EIGEN)
+else
+ifdef SUPERLU
+LIBS += -L$(SUPERLU)/lib -Wl,-rpath,$(SUPERLU)/lib -lsuperlu
+else
+LIBS += -lsuperlu
+endif
+endif
 
-all: test1 test2 test3 test4 run
+all: test1 test2 test3 test4 test5 run
 
 test1: test1.cpp
 	$(CXX) $(CXXFLAGS) $(INCPATH) -o test1 test1.cpp $(LIBS)
@@ -29,11 +39,15 @@ test3: test3.cpp
 test4: test4.cpp
 	$(CXX) $(CXXFLAGS) $(INCPATH) -o test4 test4.cpp $(LIBS)
 
+test5: test5.cpp
+	$(CXX) $(CXXFLAGS) $(INCPATH) -o test5 test5.cpp $(LIBS)
+
 run: test1 test2 test3 test4
 	LD_LIBRARY_PATH=../mole_C++ ./test1
 	LD_LIBRARY_PATH=../mole_C++ ./test2
 	LD_LIBRARY_PATH=../mole_C++ ./test3
 	LD_LIBRARY_PATH=../mole_C++ ./test4
+	LD_LIBRARY_PATH=../mole_C++ ./test5
 
 clean:
-	rm -f test1 test2 test3 test4
+	rm -f test1 test2 test3 test4 test5
