### 2024-06-01 jcorbino/mole#12 @jcorbino
@victorapm I've added a section to the README file to reference the examples and tests. Here's a screenshot:
![Screenshot from 2024-05-31 18-22-17](https://github.com/jcorbino/mole/assets/24197859/12ad7a58-c144-41bf-ad93-f62a843ab638)
