### 2024-06-01 jcorbino/mole#12 @jcorbino
@jakelangham I created a folder analogous to tests_C++ with five tests for the MATLAB version of MOLE. In the README file, I invite the users to first run such tests by executing the script `tests_MATLAB/run_tests.m`. As per your suggestion, the fifth test performs an accuracy test.

Thanks for the suggestions!
