### 2024-06-01 jcorbino/mole#16 @jcorbino
@jakelangham Thank you for bringing this issue to my attention. Improving the documentation is indeed a high priority for me. In the meantime, to address your concerns and meet the journal's requirements, I have:

1.      Expanded the C++ Doxygen documentation by adding more detailed information to the files, classes, and functions.

2.      Created a quick guide specifically for C++ users.

I am committed to enhancing the scope and quality of MOLE's documentation, and I'll get additional support this Fall to make further improvements.
