### 2024-06-13 jcorbino/mole#12 @jakelangham
Hi, @jcorbino. This seems like good progress to me. The tests seem sound and pass when I run them.

However, I am not convinced they fully address what I was getting at in https://github.com/jcorbino/mole/issues/12#issuecomment-1956529354 What I am concerned about is that they only verify the correctness of the simplest subset of the operators that may be constructed using MOLE.

For example, what I can see from the source is that the higher-order operators (which become increasingly complex) are dealt with via switch case statements. Therefore, the fact that the tests pass for the k = 2 cases can tell me nothing about whether there is an accidental typo in one of the higher order cases.

The guideline that I am trying to follow from the JOSS docs is that the tests should be able to 'objectively check the expected functionality of the software'. In my view, this means the tests should at least cover the available discretisation orders, as well as some of the different grids advertised, for each of the core differential operators supported. I am envisaging that this could be achieved by augmenting the existing tests to include loops that iterate over these different choices.
