### 2024-06-13 jcorbino/mole#16 @jakelangham
Hi @jcorbino, I appreciate the work you've put into this. The pdf guide is especially welcome.

The API docs could be improved quite a lot still, so consider continuing to add to them in the future, with more high-level information about how the underlying methods work. Nevertheless, I think there is enough there for new users to satisfy the requirements for the review.
