### 2024-06-13 #6288 @jakelangham
@jcorbino @jedbrown I think there is still more to do regarding testing/functionality (cf https://github.com/jcorbino/mole/issues/12) but upon resolution of this I would gladly recommend acceptance.
