### 2024-06-15 b8c17e1
Incorporated suggestions by Jake Langham - Expanded tests to iterate over different orders of accuracy
