+++ tests_MATLAB/test2.m
@@ -2,19 +2,22 @@
 
 addpath('../mole_MATLAB')
 
-k = 2;
-m = 2 * k + 1;
-dx = 1;
+ks = [2, 4, 6, 8];  % Different orders of accuracy
 tol = 1e-12;
 
-G = grad(k, m, dx);
-
-field = ones(m + 2, 1);
-
-sol = G * field;
-
-if (norm(sol) < tol)
-    fprintf("Test PASSED!\n");
-else
-    fprintf("Test FAILED!");
+for k = ks
+    m = 2 * k + 1;
+    dx = 1 / m;
+    
+    G = grad(k, m, dx);
+    
+    field = ones(m + 2, 1);
+    
+    sol = G * field;
+    
+    if (norm(sol) > tol)
+        fprintf("Test FAILED!\n");
+    end
 end
+
+fprintf("Test PASSED!\n");
