+++ tests_MATLAB/test5.m
@@ -1,51 +1,54 @@
-% Poisson test
+% Poisson accuracy test
 
 addpath('../mole_MATLAB')
 
 west = 0;  % Domain's limits
 east = 1;
 
-k = 2;  % Operator's order of accuracy
-grid_sizes = [10, 20, 40, 80];  % Different grid sizes to test
+ks = [2, 4, 6];  % Different orders of accuracy
+grid_sizes = [20, 40];  % Grid sizes to test
 
-errors = zeros(size(grid_sizes));
+for k = ks
 
-for i = 1:numel(grid_sizes)
-    m = grid_sizes(i);  % Number of cells
-    dx = (east - west) / m;  % Step length
+    errors = zeros(size(grid_sizes));
     
-    L = lap(k, m, dx);  % 1D Mimetic laplacian operator
+    for i = 1:numel(grid_sizes)
+        m = grid_sizes(i);  % Number of cells
+        dx = (east - west) / m;  % Step length
 
-    % Impose Robin BC on laplacian operator
-    a = 1;
-    b = 1;
-    L = L + robinBC(k, m, dx, a, b);
+        L = lap(k, m, dx);  % 1D Mimetic Laplacian operator
 
-    % 1D Staggered grid
-    grid = [west west+dx/2 : dx : east-dx/2 east];
+        % Impose Robin BC on Laplacian operator
+        a = 1;
+        b = 1;
+        L = L + robinBC(k, m, dx, a, b);
 
-    % RHS
-    U = exp(grid)';
-    U(1) = 0;  % West BC
-    U(end) = 2*exp(1);  % East BC
+        % 1D Staggered grid
+        grid = [west west+dx/2 : dx : east-dx/2 east];
 
-    % Solve a linear system of equations
-    computed_solution = L\U;
-    
-    % Compute error using L2 norm
-    analytical_solution = exp(grid);
-    errors(i) = max(abs(computed_solution' - analytical_solution));
-end
+        % RHS
+        U = exp(grid)';
+        U(1) = 0;  % West BC
+        U(end) = 2*exp(1);  % East BC
 
-% Compute order of accuracy
-order = zeros(numel(errors) - 1, 1);
-for i = 1:numel(errors) - 1
-    order(i) = log2(errors(i) / errors(i + 1));
+        % Solve a linear system of equations
+        computed_solution = L\U;
 
-    if order(i) - k < -0.5
-       fprintf("Test FAILED!");
-       return
+        % Compute error using L2 norm
+        analytical_solution = exp(grid);
+        errors(i) = max(abs(computed_solution' - analytical_solution));
+    end
+
+    % Compute order of accuracy
+    order = zeros(numel(errors) - 1, 1);
+    for i = 1:numel(errors) - 1
+        order(i) = log2(errors(i) / errors(i + 1));
+
+        if order(i) - k < -0.5
+            fprintf("Test FAILED for k = %d!\n", k);
+            return
+        end
     end
 end
 
-fprintf("Test PASSED!");
+fprintf("Test PASSED!\n");
