+++ tests_MATLAB/run_tests.m
@@ -1,7 +1,7 @@
 % Main script to run all the tests
 
 clc
-clear;
+clear all;
 close all;
 
 disp('Running: Nullity test of Divergence operator...');
@@ -16,5 +16,5 @@ run('test3.m');
 disp('Running: Energy test (Schrödinger equation)...');
 run('test4.m');
 
-disp('Running: Elliptic test (Poisson equation)...');
+disp('Running: Accuracy test (Poisson equation)...');
 run('test5.m');
