### 2024-06-15 jcorbino/mole#12 @jcorbino
Thank you, @jakelangham , for your valuable feedback! I have incorporated your suggestion of augmenting the existing tests to include loops that iterate over different orders of accuracy. You can view the changes in this [commit](https://github.com/jcorbino/mole/commit/b8c17e131e002fba6e2223833c396d44123a5242). Thanks again for bringing this to my attention! I hope this will be sufficient for the JOSS submission.

As a side note, we have just been informed that we received a grant from the NSF to expand the library. I will ensure that documentation and testing are top priorities on our to-do list.
