+++ tests_C++/test2.cpp
@@ -15,7 +15,7 @@ void run_nullity_test(int k, Real tol) {
   vec sol = G * field;
 
   if (norm(sol) > tol) {
-    std::cout << "\033[1;31mTest FAILED!\033[0m\n";
+    cout << "\033[1;31mTest FAILED!\033[0m\n";
     exit(1);
   }
 }
@@ -26,7 +26,7 @@ int main() {
   for (int k : {2, 4, 6})
     run_nullity_test(k, tol);
 
-  std::cout << "\033[1;32mTest PASSED!\033[0m\n";
+  cout << "\033[1;32mTest PASSED!\033[0m\n";
 
   return 0;
 }
