### 2024-06-19 jcorbino/mole#12 @jakelangham
Hi, @jcorbino. First of all, congratulations on your grant!

I'm glad to see the iteration over the operator order. Can I ask why you are not also including the 8th order operators? What about testing these operators on different grids etc? (Maybe it is just a case of moving some of the examples over into the test directory and tidying up?)
