+++ tests_C++/test2.cpp
@@ -16,14 +16,15 @@ void run_nullity_test(int k, Real tol) {
 
   if (norm(sol) > tol) {
     cout << "\033[1;31mTest FAILED!\033[0m\n";
+    cout << norm(sol);
     exit(1);
   }
 }
 
 int main() {
-  Real tol = 1e-12;
+  Real tol = 1e-10;
 
-  for (int k : {2, 4, 6})
+  for (int k : {2, 4, 6, 8})
     run_nullity_test(k, tol);
 
   cout << "\033[1;32mTest PASSED!\033[0m\n";
