### 2024-06-22 jcorbino/mole#12 @jcorbino
@jakelangham  Thank you!

I have added 8th-order operators to the pertinent tests (the divergence should not be requested for orders > 6 due to negative weights and the absence of a norm). Regarding the different grids: Our library, for the moment, does not provide grid generation routines. We have a couple of scripts for 2D grid generation using transfinite interpolation and elliptic grid generation. However, we do not advertise this since, for now, the user is responsible for providing the grids. In the tests, I evaluate the operators for different grid resolutions. The point of the tests is to ensure that:

1. In the C++ case, all dependencies are installed and visible to MOLE.

2. The core operators (both flavors) are correct and provide the advertised order of accuracy.

**We are fully aware of the potential for further expansion and improvement in our testing procedures.** We are committed to working on this and enhancing our documentation. However, I would like to know if the current level of testing meets the review criteria for JOSS.
