### 2024-06-22 jcorbino/mole#12 @jakelangham
@jcorbino If the divergence should not be used at 8th order then I would suggest adding this to your documentation and/or providing some indication of this via the way the underlying code works.

But in any case, I am running out of things to pick holes in here so that probably means it's ready to go back to the editor.
