+++ JOSS/paper.md
@@ -27,7 +27,7 @@ bibliography: paper.bib
 
 MOLE is an open-source library that implements high-order mimetic operators. It provides discrete analogs of the most common vector calculus operators: divergence, gradient, curl, and Laplacian. These operators act on functions discretized over staggered grids (uniform, nonuniform, and curvilinear), and they satisfy local and global conservation laws [@R202203; @R202305]. MOLE's operators can be utilized to develop code for solving partial differential equations (PDEs).
 
-The mathematics are based on the work of [@Corbino]. In addition, the user may find useful previous publications such as [@Castillo], in which similar operators are derived using a matrix analysis approach.
+The mathematics are based on the work of @Corbino. In addition, the user may find useful previous publications such as @Castillo, in which similar operators are derived using a matrix analysis approach.
 
 # Mimetic operators
 
@@ -35,7 +35,7 @@ Mimetic operators, divergence (**D**), gradient (**G**), curl (**C**), and Lapla
 
 The basis of higher-dimensional operators, as well as more sophisticated operators such as the Laplacian or the biharmonic operator are the one-dimensional mimetic **G** and **D** operators, together with high-order mimetic interpolation operators [@R202202], which are also contained in the library. These finite-dimensional operators can be reused throughout the mathematical model and they provide a higher level of abstraction at the time of solving PDEs.
 
-These operators, have been used to write codes to solve PDEs of different types [@Jared; @Bazan; @Boada; @Boada2; @Rojas; @Jorge; @Josep; @Abouali]. For an overview of mimetic methods of different types see the book by Castillo and Miranda and the references there in [@Miranda].
+These operators have been used to write codes to solve PDEs of different types [@Jared; @Bazan; @Boada; @Boada2; @Rojas; @Jorge; @Josep; @Abouali]. For an overview of mimetic methods of different types see the book by Castillo and Miranda and the references therein [@Miranda].
 
 # Statement of need
 
@@ -43,11 +43,11 @@ Implementing mimetic operators, particularly in three dimensions, presents signi
 
 # State of the field
 
-A previous library [@MTK] was developed to implement the mimetic operators presented in [@Castillo]. This library was only capable of handling dense matrices so it was limited to solve small problems hence its development was stopped. MOLE implements the operators presented in the Corbino and Castillo paper [@Corbino]. These  operators are optimal from the number of points in each stencil and produce more accurate results. MOLE deals with sparse matrices efficiently and is capable of solving problems with millions of cells. To the best of the authors' knowledge, there are no other libraries that implement mimetic methods as the ones presented in this paper.
+A previous library [@MTK] was developed to implement the mimetic operators presented in @Castillo. This library was only capable of handling dense matrices so it was limited to solve small problems hence its development was stopped. MOLE implements the operators presented in @Corbino. These operators are optimal from the number of points in each stencil and produce more accurate results. MOLE deals with sparse matrices efficiently and is capable of solving problems with millions of cells. To the best of the authors' knowledge, there are no other libraries that implement mimetic methods as the ones presented in this paper.
 
 # The library
 
-MOLE was designed to be an intuitive software package to construct mimetic operators based on [@Corbino] method. MOLE is implemented in C++ and in MATLAB scripting language (these are two independent flavors) and every single function in MOLE returns a sparse matrix of the requested mimetic operator. For information on the installation or usage of the library, please read the [documentation](https://github.com/jcorbino/mole/blob/master/README.md) included in the repository.
[…]
