### 2024-07-16 #6288 @jedbrown
Thanks for your patience. Please have a look at https://github.com/jcorbino/mole/pull/18 and merge if you feel it is acceptable. Then proceed to:
