### 2024-07-16 #6288 @jcorbino
Not a problem! Thank you @jedbrown 👍🏻

I just merged the PR, and here's the requested info:

- https://github.com/jcorbino/mole/releases/tag/v1.0
- https://doi.org/10.5281/zenodo.12752946
