### 2024-07-17 #6288 @jedbrown
Thanks. I just want to confirm that you are intentionally clobbering the previous `v1.0` tag (there were previous tags up to `v5.0`) with this new versioning scheme. If so, I think you should update your `CITATION.cff` (which currently has `version: v2.0` from last year). An alternative is to call this `v6.0` than to start over with a new versioning scheme.
