### 2024-07-17 0d03c22
Update CITATION.cff as recommended by Jed Brown
