+++ CITATION.cff
@@ -5,7 +5,7 @@ authors:
     given-names: "Corbino"
     orcid: "https://orcid.org/0000-0002-2638-9216"
 title: "MOLE: Mimetic Operators Library Enhanced"
-version: v2.0
-doi: 10.5281/zenodo.7996615
-date-released: 2023-06-01
+version: v1.0
+doi: 10.5281/zenodo.12752946
+date-released: 2024-07-17
 url: "https://github.com/jcorbino/mole"
