### 2024-07-17 #6288 @jcorbino
Thank you, @jedbrown - I completely overlooked that. I've now updated the `CITATION.cff` with the new version and DOI. I prefer keeping this as version 1.0. Thanks again for pointing that out! 👍🏻
