### 2024-07-17 #6288 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
