### 2024-07-17 da34873
minor changes in paper.md
