+++ JOSS/paper.bib
@@ -113,7 +113,7 @@
   url = {https://joss.theoj.org/papers/10.21105/joss.00026},
   author = {{Sanderson}, Conrad and {Curtin}, Ryan},
   doi = {10.21105/joss.00026},
-  journal = {Open Source Software},
+  journal = {Journal of Open Source Software},
   month = July,
   title = {Armadillo: a template-based {C++} library for linear algebra},
   year = 2016,
