### 2024-07-17 5103d48
Update CITATION.cff
