### 2024-07-17 #6288 @danielskatz
Congratulations to @jcorbino (Johnny Corbino) and co-authors on your publication!!

JOSS depends on volunteers and wouldn't be successful without you.
